REFACTOR: 구조 리팩토링으로 인한 Test코드 수정(#52)
